fwdstatus: reject stale sampler data (#10842)
Invalidate CPU windows when the newest sample is older than two sampler intervals, and mark worker windows spanning held telemetry as unavailable. Detect USER_HZ and page size at runtime, label the scheduler fallback as approximate, and annotate the busy-poll CPU false-positive.

+++ pkg/fwdstatus/ticks_overflow_4909_test.go
@@ -6,20 +6,19 @@ import (
 )
 
 // TestTicksToNanosNoOverflow pins #4909: converting a large scheduler-tick
-// count to nanoseconds must not overflow. The pre-fix `ticks * 1e9 / userHZ`
-// wraps uint64 once `ticks * 1e9` exceeds 2^64 — at ~1.845e10 ticks (~33 days
-// of busy CPU summed across 64 cores at 100 Hz) — corrupting the CPU windows
-// used to diagnose saturation. The divide-before-multiply form is exact.
+// count to nanoseconds must not overflow. Multiplying ticks by 1e9 before
+// dividing by the runtime USER_HZ wraps uint64 once ticks exceed
+// approximately 1.845e10, corrupting the CPU windows used to diagnose
+// saturation. Divide-before-multiply preserves the exact conversion.
 //
-// RED on revert: restore `ticks * 1e9 / userHZ` and the wrap/near-wrap cases
-// below produce a wrong (wrapped) value, mismatching the big.Int oracle.
+// RED on revert: restore the overflowing intermediate and the wrap/near-wrap
+// cases below produce a wrong result, mismatching the big.Int oracle.
 func TestTicksToNanosNoOverflow(t *testing.T) {
 	// Old intermediate wrap point: 2^64 / 1e9 ≈ 1.8446744073e10 ticks.
 	const wrap = ^uint64(0) / 1_000_000_000
 
-	// Upper bound of the inputs: the true result must still fit uint64 (the
-	// function returns uint64 nanoseconds). 1<<40 ticks ≈ 1.1e19 ns fits;
-	// anything much larger is unphysical (centuries of CPU at 100 Hz).
+	// Upper bound of the inputs: the exact runtime-HZ conversion must still
+	// fit uint64 for the tested Linux scheduler rate.
 	cases := []uint64{
 		0, 1, 99, 100, 101, 12_345,
 		wrap - 1, wrap, wrap + 1, wrap + 1000,
@@ -27,7 +26,7 @@ func TestTicksToNanosNoOverflow(t *testing.T) {
 	}
 
 	bilNano := big.NewInt(1_000_000_000)
-	bilHZ := big.NewInt(userHZ)
+	bilHZ := big.NewInt(int64(userHZ))
 	for _, ticks := range cases {
 		got := ticksToNanos(ticks)
